Finished initial rewrite for the portions of NoAccountActivity dealing with the new message-passing design outlined in https://github.com/afwang/ogame-android-app/issues/21. This redesign is to remove the long-running service from the background.

+++ app/src/main/java/com/wikaba/ogapp/events/OnAgentUpdateEvent.java
@@ -21,7 +21,7 @@ package com.wikaba.ogapp.events;
  * Created by afwang on 9/7/15.
  */
 public class OnAgentUpdateEvent {
-	private int agentKey;
+	private long agentKey;
 	private boolean updateSucceeded;
 
 	public OnAgentUpdateEvent() {
@@ -34,7 +34,7 @@ public class OnAgentUpdateEvent {
 		updateSucceeded = true;
 	}
 
-	public OnAgentUpdateEvent(int agentManagerKey, boolean wasUpdateSuccessful) {
+	public OnAgentUpdateEvent(long agentManagerKey, boolean wasUpdateSuccessful) {
 		agentKey = agentManagerKey;
 		updateSucceeded = wasUpdateSuccessful;
 	}
@@ -43,7 +43,7 @@ public class OnAgentUpdateEvent {
 		return updateSucceeded;
 	}
 
-	public int getAgentManagerKey() {
+	public long getAgentManagerKey() {
 		return agentKey;
 	}
 }
